refactor: extract 'ManuallyReportedException' to :common
The class is a pure RuntimeException subclass with no Android dependencies,
so it lives in :common rather than :common:android. Removes one more
dependency on :AnkiDroid for feature module extraction.
