+++ AnkiDroid/src/main/java/com/ichi2/anki/snackbar/Snackbars.kt
@@ -29,7 +29,7 @@ import com.google.android.material.snackbar.onAttachedToWindow2
 import com.ichi2.anki.BuildConfig
 import com.ichi2.anki.R
 import com.ichi2.anki.common.crashreporting.CrashReportService
-import com.ichi2.anki.exception.ManuallyReportedException
+import com.ichi2.anki.common.exception.ManuallyReportedException
 import com.ichi2.anki.showThemedToast
 import timber.log.Timber
