+++ AnkiDroid/src/main/java/com/ichi2/anki/workarounds/AppLoadedFromBackupWorkaround.kt
@@ -22,7 +22,7 @@ import android.os.Process
 import com.ichi2.anki.AnkiDroidApp
 import com.ichi2.anki.R
 import com.ichi2.anki.common.crashreporting.CrashReportService
-import com.ichi2.anki.exception.ManuallyReportedException
+import com.ichi2.anki.common.exception.ManuallyReportedException
 import com.ichi2.anki.showThemedToast
 import com.ichi2.themes.Themes
 import timber.log.Timber
